Fixed missing constant import, split testing data and moved createResponses to a shared function
